% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcD.R
\name{calcD}
\alias{calcD}
\title{Calculate distance metric between rct3 prediction and assumed value}
\usage{
calcD(rct3Obj = NULL, A = NULL)
}
\arguments{
\item{rct3Obj}{results object of class rct3}

\item{vector.}{Assumed numbers to compare to rct prediction}
}
\value{
vector
}
\description{
The RCT3 analysis gives a year‐class strength prediction based 
  on the survey information, and the standard error associated with the 
  prediction. The difference between the assumed size of the recruiting 
  year class in spring (before autumn surveys are available) and the RCT3 
  year‐class strength estimates based on summer surveys, scaled to the 
  internal standard error calculated by RCT3, is D = (R-A)/S, where R is the 
  log Weighted Average Prediction (WAP) from RCT3, A is the assumed 
  year‐class strength in spring assessment report, and S is the internal 
  standard error from RCT3 (ICES, 2008).
}
\examples{

data(pok)
fmla <- formula(age4 ~ ibtsq3)
res <- rct3(formula = fmla, data = pok)
calcD(rct3Obj = res, A = log(83000))

}
\references{
ICES. 2008. Report of the Ad hoc Group on Criteria for Reopening Fisheries 
  Advice (AGCREFA), 20–22 August 2008, Copenhagen, Denmark. 
  ICES CM 2008/ACOM:60. 30 pp.
}
