% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.rct3.R
\name{plot.rct3}
\alias{plot.rct3}
\title{Plot rct3 regression fits by index}
\usage{
\method{plot}{rct3}(rct3Obj, predYr = NULL, indices = NULL, ylimConst = TRUE)
}
\arguments{
\item{rct3Obj}{rct object}

\item{predYr}{prediciont year}

\item{indices}{Indices to plot (Default = NULL plots all)}

\item{ylimConst}{logical. Use same y-lmits for all indices.}
}
\value{
plot
}
\description{
Produces a plot of the rct3 model term fits for a single year.
  Regression lines for each term are a black dashed line, mean  
  (black dashed regression line), including predicted value and confidence interval (blue).
}
\examples{
data(pok)
head(pok)
fmla <- formula(age4 ~ ibtsq3)
pok_rct3 <- rct3(formula = fmla, data = pok, range = nrow(pok), old = FALSE)
plot(pok_rct3)


data(recdata)
head(recdata)
formula <- recruitment ~ NT1 + NT2 + NT3 +
  NAK1 + NAK2 + NAK3 +
  RT1 + RT2 + RT3 +
  ECO1 + ECO2 + ECO3

my_rct3 <- rct3(formula = formula, data = recdata, predictions = 2012:2017, 
  shrink = TRUE)
# plot(my_rct3) # produces error since no prediction year is defined
plot(rct3Obj = my_rct3, predYr = 2012)
plot(rct3Obj = my_rct3, predYr = 2012, ylimConst = F) # variable ylim



}
