% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pok.R
\docType{data}
\name{pok}
\alias{pok}
\title{Recruitment and survey index data for North Sea saith (pok.27.3a46)}
\format{Data frame containing 14 columns:
\tabular{ll}{
  \code{yearclass} \tab the yearclass\cr
  \code{age4} \tab the recruiment (age 4) for that yearclass\cr
  \code{ibtsq3} \tab The age 4 survey index from the 'IBTS' quarter 3 survey
}}
\usage{
pok
}
\description{
data.frame containing recruitment (age 4) and survey indices from the 
IBTS quarter 3 survey.
}
\examples{
data(pok)
head(pok)
}
\seealso{
\code{\link{rct3}}  run a calibrated regression to predict rectruitment.

\code{\link{rct3-package}} gives an overview of the package.
}
