% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct3-package.R
\docType{package}
\name{rct3-package}
\alias{rct3-package}
\title{Function to run a calibrated regression to predict recruitment}
\description{
Function to run a calibrated regression to predict recruitment using the
method decribed by Shepherd
}
\details{
\tabular{ll}{
  \code{\link{rct3}} \tab run a calibrated regression to predict rectruitment\cr
  \code{\link{print.rct3}}  \tab print a rct3 object\cr
  \code{\link{summary.rct3}} \tab summarise a rct3 object
}
}
\author{
Colin Millar.
}
