% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shep.R
\docType{data}
\name{shep}
\alias{shep}
\title{Recruitment and survey index data from Shepherd (1997)}
\format{Data frame containing 14 columns:
\tabular{ll}{
  \code{yearclass} \tab the yearclass\cr
  \code{age1} \tab the recruiment (age 1) for that yearclass\cr
  \code{ssoct0} \tab The age 0 survey index from the October 'NWGFS' survey\cr
  \code{ssjun1} \tab The age 1 survey index from June 'NWGFS' survey\cr
  \code{...} \tab and so on
}}
\usage{
shep
}
\description{
data.frame containing recruitment (age 1) and survey indices from several
surveys over ages 0 to 2. See Shepherd (1997) for details.
}
\examples{
data(shep)
head(shep)
}
\seealso{
\code{\link{rct3}}  run a calibrated regression to predict rectruitment.

\code{\link{rct3-package}} gives an overview of the package.
}
